# graph/internal/set: put Nodes methods on the pointer type