--- a/graph/internal/set/same.go
+++ b/graph/internal/set/same.go
@@ -13,8 +13,8 @@ import "unsafe"
 // hash maps are passed as a pointer to a runtime Hmap struct. A map is
 // not seen by the runtime as a pointer though, so we use unsafe to get
 // the maps' pointer values to compare.
-func same(a, b Nodes) bool {
-	return *(*uintptr)(unsafe.Pointer(&a)) == *(*uintptr)(unsafe.Pointer(&b))
+func same(a, b *Nodes) bool {
+	return a == b || *(*uintptr)(unsafe.Pointer(a)) == *(*uintptr)(unsafe.Pointer(b))
 }
 
 // intsSame determines whether two sets are backed by the same store. In the