--- a/graph/internal/set/same_appengine.go
+++ b/graph/internal/set/same_appengine.go
@@ -13,8 +13,8 @@ import "reflect"
 // hash maps are passed as a pointer to a runtime Hmap struct. A map is
 // not seen by the runtime as a pointer though, so we use reflect to get
 // the maps' pointer values to compare.
-func same(a, b Nodes) bool {
-	return reflect.ValueOf(a).Pointer() == reflect.ValueOf(b).Pointer()
+func same(a, b *Nodes) bool {
+	return a == b || reflect.ValueOf(*a).Pointer() == reflect.ValueOf(*b).Pointer()
 }
 
 // intsSame determines whether two sets are backed by the same store. In the