--- a/graph/internal/set/set_test.go
+++ b/graph/internal/set/set_test.go
@@ -13,11 +13,6 @@ type node int64
 
 func (n node) ID() int64 { return int64(n) }
 
-// count reports the number of elements stored in the node set.
-func (s Nodes) count() int {
-	return len(s)
-}
-
 // TestSame tests the assumption that pointer equality via unsafe conversion
 // of a map[int]struct{} to uintptr is a valid test for perfect identity between
 // set values. If any of the tests in TestSame fail, the package is broken and same
@@ -318,9 +313,6 @@ func TestSameNodes(t *testing.T) {
 	if !same(nil, nil) {
 		t.Error("nil sets test as not same.")
 	}
-	if same(b, nil) {
-		t.Error("nil and empty sets test as same.")
-	}
 }
 
 func TestAddNodes(t *testing.T) {
@@ -329,7 +321,7 @@ func TestAddNodes(t *testing.T) {
 		t.Fatal("Set cannot be created successfully")
 	}
 
-	if s.count() != 0 {
+	if s.Count() != 0 {
 		t.Error("Set somehow contains new elements upon creation")
 	}
 
@@ -337,7 +329,7 @@ func TestAddNodes(t *testing.T) {
 	s.Add(node(3))
 	s.Add(node(5))
 
-	if s.count() != 3 {
+	if s.Count() != 3 {
 		t.Error("Incorrect number of set elements after adding")
 	}
 
@@ -347,9 +339,9 @@ func TestAddNodes(t *testing.T) {
 
 	s.Add(node(1))
 
-	if s.count() > 3 {
+	if s.Count() > 3 {
 		t.Error("Set double-adds element (element not unique)")
-	} else if s.count() < 3 {
+	} else if s.Count() < 3 {
 		t.Error("Set double-add lowered len")
 	}
 
@@ -361,7 +353,7 @@ func TestAddNodes(t *testing.T) {
 		t.Error("Set removes element on double-add")
 	}
 
-	for e, n := range s {
+	for e, n := range *s {
 		if e != n.ID() {
 			t.Errorf("Element ID did not match key: %d != %d", e, n.ID())
 		}
@@ -377,7 +369,7 @@ func TestRemoveNodes(t *testing.T) {
 
 	s.Remove(node(1))
 
-	if s.count() != 2 {
+	if s.Count() != 2 {
 		t.Error("Incorrect number of set elements after removing an element")
 	}
 
@@ -391,13 +383,13 @@ func TestRemoveNodes(t *testing.T) {
 
 	s.Remove(node(1))
 
-	if s.count() != 2 || s.Has(node(1)) {
+	if s.Count() != 2 || s.Has(node(1)) {
 		t.Error("Double set remove does something strange")
 	}
 
 	s.Add(node(1))
 
-	if s.count() != 3 || !s.Has(node(1)) {
+	if s.Count() != 3 || !s.Has(node(1)) {
 		t.Error("Cannot add element after removal")
 	}
 }
@@ -411,7 +403,7 @@ func TestClearNodes(t *testing.T) {
 
 	s.clear()
 
-	if s.count() != 0 {
+	if s.Count() != 0 {
 		t.Error("clear did not properly reset set to size 0")
 	}
 }
@@ -463,7 +455,7 @@ func TestCopyNodes(t *testing.T) {
 	a.Add(node(2))
 	a.Add(node(3))
 
-	b.Copy(a)
+	b.Clone(a)
 
 	if !Equal(a, b) {
 		t.Fatalf("Two sets not equal after copy: %v != %v", a, b)
@@ -475,7 +467,7 @@ func TestCopyNodes(t *testing.T) {
 		t.Errorf("Mutating one set mutated another after copy: %v == %v", a, b)
 	}
 
-	b = b.Copy(a) // Assignment is needed because b is not empty.
+	b.Clone(a)
 
 	if !Equal(a, b) {
 		t.Fatalf("Two sets not equal after second copy: %v != %v", a, b)
@@ -488,9 +480,9 @@ func TestSelfCopyNodes(t *testing.T) {
 	a.Add(node(1))
 	a.Add(node(2))
 
-	a.Copy(a)
+	a.Clone(a)
 
-	if a.count() != 2 {
+	if a.Count() != 2 {
 		t.Error("Something strange happened when copying into self")
 	}
 }
@@ -508,7 +500,7 @@ func TestUnionSameNodes(t *testing.T) {
 
 	c.Union(a, b)
 
-	if c.count() != 2 {
+	if c.Count() != 2 {
 		t.Error("Union of same sets yields set with wrong len")
 	}
 
@@ -516,8 +508,8 @@ func TestUnionSameNodes(t *testing.T) {
 		t.Error("Union of same sets yields wrong elements")
 	}
 
-	for i, s := range []Nodes{a, b, c} {
-		for e, n := range s {
+	for i, s := range []*Nodes{a, b, c} {
+		for e, n := range *s {
 			if e != n.ID() {
 				t.Errorf("Element ID did not match key in s%d: %d != %d", i+1, e, n.ID())
 			}
@@ -537,7 +529,7 @@ func TestUnionDiffNodes(t *testing.T) {
 
 	c.Union(a, b)
 
-	if c.count() != 3 {
+	if c.Count() != 3 {
 		t.Error("Union of different sets yields set with wrong len")
 	}
 
@@ -545,23 +537,23 @@ func TestUnionDiffNodes(t *testing.T) {
 		t.Error("Union of different sets yields set with wrong elements")
 	}
 
-	if a.Has(node(3)) || !a.Has(node(2)) || !a.Has(node(1)) || a.count() != 2 {
+	if a.Has(node(3)) || !a.Has(node(2)) || !a.Has(node(1)) || a.Count() != 2 {
 		t.Error("Union of sets mutates non-destination set (argument 1)")
 	}
 
-	if !b.Has(node(3)) || b.Has(node(1)) || b.Has(node(2)) || b.count() != 1 {
+	if !b.Has(node(3)) || b.Has(node(1)) || b.Has(node(2)) || b.Count() != 1 {
 		t.Error("Union of sets mutates non-destination set (argument 2)")
 	}
 
-	for i, s := range []Nodes{a, b, c} {
-		for e, n := range s {
+	for i, s := range []*Nodes{a, b, c} {
+		for e, n := range *s {
 			if e != n.ID() {
 				t.Errorf("Element ID did not match key in s%d: %d != %d", i+1, e, n.ID())
 			}
 		}
 	}
 
-	c = c.Union(a, a) // Assignment is necessary because c is not empty.
+	c.Union(a, a)
 	if !reflect.DeepEqual(c, a) {
 		t.Errorf("Union of equal sets not equal to sets: %v != %v", c, a)
 	}
@@ -580,7 +572,7 @@ func TestUnionOverlappingNodes(t *testing.T) {
 
 	c.Union(a, b)
 
-	if c.count() != 3 {
+	if c.Count() != 3 {
 		t.Error("Union of overlapping sets yields set with wrong len")
 	}
 
@@ -588,23 +580,23 @@ func TestUnionOverlappingNodes(t *testing.T) {
 		t.Error("Union of overlapping sets yields set with wrong elements")
 	}
 
-	if a.Has(node(3)) || !a.Has(node(2)) || !a.Has(node(1)) || a.count() != 2 {
+	if a.Has(node(3)) || !a.Has(node(2)) || !a.Has(node(1)) || a.Count() != 2 {
 		t.Error("Union of sets mutates non-destination set (argument 1)")
 	}
 
-	if !b.Has(node(3)) || b.Has(node(1)) || !b.Has(node(2)) || b.count() != 2 {
+	if !b.Has(node(3)) || b.Has(node(1)) || !b.Has(node(2)) || b.Count() != 2 {
 		t.Error("Union of sets mutates non-destination set (argument 2)")
 	}
 
-	for i, s := range []Nodes{a, b, c} {
-		for e, n := range s {
+	for i, s := range []*Nodes{a, b, c} {
+		for e, n := range *s {
 			if e != n.ID() {
 				t.Errorf("Element ID did not match key in s%d: %d != %d", i+1, e, n.ID())
 			}
 		}
 	}
 
-	c = c.Intersect(a, a) // Assignment is necessary because c is not empty.
+	c.Intersect(a, a)
 	if !reflect.DeepEqual(c, a) {
 		t.Errorf("Intersection of equal sets not equal to sets: %v != %v", c, a)
 	}
@@ -623,7 +615,7 @@ func TestIntersectSameNodes(t *testing.T) {
 
 	c.Intersect(a, b)
 
-	if card := c.count(); card != 2 {
+	if card := c.Count(); card != 2 {
 		t.Errorf("Intersection of identical sets yields set of wrong len %d", card)
 	}
 
@@ -631,8 +623,8 @@ func TestIntersectSameNodes(t *testing.T) {
 		t.Error("Intersection of identical sets yields set of wrong elements")
 	}
 
-	for i, s := range []Nodes{a, b, c} {
-		for e, n := range s {
+	for i, s := range []*Nodes{a, b, c} {
+		for e, n := range *s {
 			if e != n.ID() {
 				t.Errorf("Element ID did not match key in s%d: %d != %d", i+1, e, n.ID())
 			}
@@ -653,20 +645,20 @@ func TestIntersectDiffNodes(t *testing.T) {
 
 	c.Intersect(a, b)
 
-	if card := c.count(); card != 0 {
+	if card := c.Count(); card != 0 {
 		t.Errorf("Intersection of different yields non-empty set %d", card)
 	}
 
-	if !a.Has(node(2)) || !a.Has(node(3)) || a.Has(node(1)) || a.Has(node(4)) || a.count() != 2 {
+	if !a.Has(node(2)) || !a.Has(node(3)) || a.Has(node(1)) || a.Has(node(4)) || a.Count() != 2 {
 		t.Error("Intersection of sets mutates non-destination set (argument 1)")
 	}
 
-	if b.Has(node(2)) || b.Has(node(3)) || !b.Has(node(1)) || !b.Has(node(4)) || b.count() != 2 {
+	if b.Has(node(2)) || b.Has(node(3)) || !b.Has(node(1)) || !b.Has(node(4)) || b.Count() != 2 {
 		t.Error("Intersection of sets mutates non-destination set (argument 1)")
 	}
 
-	for i, s := range []Nodes{a, b, c} {
-		for e, n := range s {
+	for i, s := range []*Nodes{a, b, c} {
+		for e, n := range *s {
 			if e != n.ID() {
 				t.Errorf("Element ID did not match key in s%d: %d != %d", i+1, e, n.ID())
 			}
@@ -687,7 +679,7 @@ func TestIntersectOverlappingNodes(t *testing.T) {
 
 	c.Intersect(a, b)
 
-	if card := c.count(); card != 1 {
+	if card := c.Count(); card != 1 {
 		t.Errorf("Intersection of overlapping sets yields set of incorrect len %d", card)
 	}
 
@@ -695,29 +687,29 @@ func TestIntersectOverlappingNodes(t *testing.T) {
 		t.Errorf("Intersection of overlapping sets yields set with wrong element")
 	}
 
-	if !a.Has(node(2)) || !a.Has(node(3)) || a.Has(node(4)) || a.count() != 2 {
+	if !a.Has(node(2)) || !a.Has(node(3)) || a.Has(node(4)) || a.Count() != 2 {
 		t.Error("Intersection of sets mutates non-destination set (argument 1)")
 	}
 
-	if b.Has(node(2)) || !b.Has(node(3)) || !b.Has(node(4)) || b.count() != 2 {
+	if b.Has(node(2)) || !b.Has(node(3)) || !b.Has(node(4)) || b.Count() != 2 {
 		t.Error("Intersection of sets mutates non-destination set (argument 1)")
 	}
 
-	for i, s := range []Nodes{a, b, c} {
-		for e, n := range s {
+	for i, s := range []*Nodes{a, b, c} {
+		for e, n := range *s {
 			if e != n.ID() {
 				t.Errorf("Element ID did not match key in s%d: %d != %d", i+1, e, n.ID())
 			}
 		}
 	}
 
-	c = c.Intersect(c, a) // Assignment is necessary because c is not empty.
+	c.Intersect(c, a)
 	want := Nodes{3: node(3)}
-	if !reflect.DeepEqual(c, want) {
-		t.Errorf("Intersection of sets with dst equal to a not equal: %v != %v", c, want)
+	if !reflect.DeepEqual(c, &want) {
+		t.Errorf("Intersection of sets with dst equal to a not equal: %v != %v", c, &want)
 	}
-	c = c.Intersect(a, c) // Assignment is necessary because c is not empty.
-	if !reflect.DeepEqual(c, want) {
-		t.Errorf("Intersection of sets with dst equal to a not equal: %v != %v", c, want)
+	c.Intersect(a, c)
+	if !reflect.DeepEqual(c, &want) {
+		t.Errorf("Intersection of sets with dst equal to a not equal: %v != %v", c, &want)
 	}
 }