--- a/graph/multi/weighted_directed_test.go
+++ b/graph/multi/weighted_directed_test.go
@@ -49,10 +49,10 @@ func weightedDirectedBuilder(nodes []graph.Node, edges []testgraph.WeightedLine,
 		e = append(e, cl)
 		dg.SetWeightedLine(cl)
 	}
-	if len(seen) != 0 {
-		n = make([]graph.Node, 0, len(seen))
+	if seen.Count() != 0 {
+		n = make([]graph.Node, 0, seen.Count())
 	}
-	for _, sn := range seen {
+	for _, sn := range *seen {
 		n = append(n, sn)
 	}
 	return dg, n, e, self, absent, true