--- a/graph/simple/weighted_directed_test.go
+++ b/graph/simple/weighted_directed_test.go
@@ -44,10 +44,10 @@ func weightedDirectedBuilder(nodes []graph.Node, edges []testgraph.WeightedLine,
 	if len(e) == 0 && len(edges) != 0 {
 		return nil, nil, nil, math.NaN(), math.NaN(), false
 	}
-	if len(seen) != 0 {
-		n = make([]graph.Node, 0, len(seen))
+	if seen.Count() != 0 {
+		n = make([]graph.Node, 0, seen.Count())
 	}
-	for _, sn := range seen {
+	for _, sn := range *seen {
 		n = append(n, sn)
 	}
 	return dg, n, e, self, absent, true