--- a/graph/topo/clique_graph.go
+++ b/graph/topo/clique_graph.go
@@ -58,7 +58,7 @@ func CliqueGraph(dst Builder, g graph.Undirected) {
 				case len(vc.Clique.nodes):
 					edgeNodes = []graph.Node{vc.Clique.nodes[0]}
 				default:
-					for _, n := range set.NewNodes().Intersect(uc.nodes, vc.nodes) {
+					for _, n := range *set.NewNodes().Intersect(uc.nodes, vc.nodes) {
 						edgeNodes = append(edgeNodes, n)
 					}
 					sort.Sort(ordered.ByID(edgeNodes))
@@ -74,7 +74,7 @@ type cliqueNodeSets map[int64][]*nodeSet
 
 type nodeSet struct {
 	Clique
-	nodes set.Nodes
+	nodes *set.Nodes
 }
 
 // Clique is a node in a clique graph.