… graph/topo/bron_kerbosch_test.go