@kortschak 2019-03-16
This has a fairly significant impact on at least one set-intense routine.

<table class='benchstat'>
<tr><th>name</th><th>old time/op</th><th>new time/op</th><th>delta</th>
<tr><td>KCliqueCommunities/simple-k=3-8</td><td>76.1µs ± 2%</td><td>88.2µs ± 2%</td><td>+15.80%</td><td>(p=0.000 n=17+20)</td>
<tr><td>KCliqueCommunities/Batagelj-Zaversnik_Graph-k=3-8</td><td>230µs ± 3%</td><td>258µs ± 3%</td><td>+12.47%</td><td>(p=0.000 n=20+20)</td>
<tr><td>KCliqueCommunities/Batagelj-Zaversnik_Graph-k=4-8</td><td>229µs ± 2%</td><td>258µs ± 2%</td><td>+12.71%</td><td>(p=0.000 n=19+20)</td>
</table>